+++ src/index.ts
@@ -1,26 +1,6 @@
 import "regenerator-runtime/runtime";
 import "whatwg-fetch";
 
-import LevelCreditProvider from "src/LevelCreditProvider";
-
-import useScore from "src/CreditAPI/score/useScore";
-import ScoreChart from "src/CreditAPI/score/ScoreChart/ScoreChart";
-import ScoreDisplay from "src/CreditAPI/score/ScoreDisplay/ScoreDisplay";
-
-import useInsights from "./CreditAPI/insights/useInsights";
-import AccountSummary from "./CreditAPI/insights/AccountSummary/AccountSummary";
-import CreditBalance from "./CreditAPI/insights/CreditBalance/CreditBalance";
-import CreditUtilization from "./CreditAPI/insights/CreditUtilization/CreditUtilization";
-import HardInquiry from "./CreditAPI/insights/HardInquiry/HardInquiry";
-
-export {
-  LevelCreditProvider,
-  useScore,
-  ScoreChart,
-  ScoreDisplay,
-  useInsights,
-  AccountSummary,
-  CreditBalance,
-  CreditUtilization,
-  HardInquiry,
-};
+export { default as LevelCreditProvider, useLevelCredit } from "src/LevelCreditProvider";
+export * from "src/CreditAPI/score";
+export * from "src/CreditAPI/insights";